move teatest to ci file
